+++ python/.gitignore
@@ -1,5 +1,3 @@
 env
+vast.egg-info
 venv
-
-# Ignore dynaconf secret files
-.secrets.*
