+++ python/README.md
@@ -9,13 +9,19 @@ Setup a virtual environment:
 ```bash
 python3 -m venv venv
 . venv/bin/activate
-pip install -r requirements.txt
+pip install -e .
+```
+
+Thereafter, you should be able to run files in the `examples` directory, e.g.:
+
+```bash
+python examples/misp.py
 ```
 
 ## Unit Tests
 
-Run all unit tests:
+Run all unit tests as follows:
 
 ```bash
-pytest tests
+pytest
 ```
