--- python/requirements.txt
@@ -1,10 +1,0 @@
-git+https://github.com/misp/misp-stix
-
-coloredlogs>=10.0
-dynaconf[yaml]
-pluggy
-pyarrow
-pytest
-threatbus
-vast-threatbus
-zmq
