+++ python/setup.cfg
@@ -1,0 +1,39 @@
+[metadata]
+name = vast
+version = 0.1.0
+author = Tenzir
+author_email = engineering@tenzir.com
+url = https://github.com/tenzir/vast
+license = BSD 3-clause
+description_file = README.md
+description = A security telemetry engine for detection and response
+long_description = file: README.md
+long_description_content_type = text/markdown; charset=UTF-8
+classifiers =
+    Development Status :: 3 - Alpha
+    Environment :: Plugins
+    License :: OSI Approved :: BSD License
+    Operating System :: MacOS :: MacOS X
+    Operating System :: POSIX :: Linux
+    Topic :: Scientific/Engineering :: Information Analysis
+    Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator
+    Topic :: Security
+    Topic :: Software Development :: Object Brokering
+    Topic :: System :: Distributed Computing
+project_urls =
+    Documentation = https://vast.io/docs
+    Source = https://github.com/tenzir/vast
+    Tracker = https://github.com/tenzir/vast/issues
+
+[options]
+packages = find:
+include_package_data = True
+install_requires =
+    coloredlogs>=10.0
+    dynaconf[yaml]
+    misp_stix_converter @ git+https://github.com/misp/misp-stix@v2.4.160
+    pandas
+    pyarrow
+    pytest
+    zmq
+python_requires = >=3.9
