+++ python/vast/backbones/__init__.py
@@ -1,0 +1,1 @@
+from .inmemory import InMemory
